v2/requirements.txt
fastapi==0.135.3
uvicorn==0.44.0
sqlalchemy==2.0.49
pydantic[email]==2.12.5
python-jose[cryptography]==3.5.0
bcrypt==5.0.0
python-multipart==0.0.26
python-dotenv==1.2.2